+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -132,7 +132,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-02-05T16:13:04.285Z")
+    updated: Optional[datetime] = parse_datetime("2026-02-05T16:44:41.846Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")