+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGQ98MQ6SWQ92C6Q9NNKS65R"
+        "01KGQB2J5G1H9N6Y8R0V7JS9DY"
     )
     r"""The unique identifier of the data source"""